simple enemy movement
-cactus data added
-randomly pick enemy type

+++ src/models/dimension.h
@@ -2,8 +2,8 @@
 #define DIMENSION_H
 
 typedef struct Dimension_s{
-	UINT16 x;
-	UINT16 y;
+	INT8 x;
+	INT8 y;
 	UINT8 width;
 	UINT8 height;
 } Dimension;
